Removed no explicit interpreter optimization
In rare cases, files that are installed via a setup.py seem to get their
executable bit set (even when they shouldn't). This causes the
emr._executable method to assume they are executable and causes it to
skip adding the interepretter to the command to be exectuted. If the
script in questions lacks (or has an incorrect) #! line then the job
will fail.

I decided the best way to fix this is safety first and always include
the interpreter.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>