The RawProtocol only handles strings
Since the JSONProtocol and PickleProtocol end up calling into libraries that can handle
any output that the mapper or reducer yields. For the RawProtocol since it was doing a
string.join, errors would occur if a non string was yielded. For example I had a reducer
that was yielding a string for the key and an int as the value.