Allow partial AMI versions.
When matching EMR AMI versions, previously an exact match was
used. However, the API accepts partial versions, e.g. instead of
major.minor.patch, provide only major or major.minor. This works when
allocating a new job flow, but subsequent calls trying to reuse that
job flow would fail to find a match because the versions didn't match
exactly.

This just turns that test into a prefix test -- if the requested
version is a prefix of the job flow's version, the job flow is
considered a match.