Recursively merge at each key/val of cluster parameters.
The function `_cluster_kwargs` builds a dict of options to submit to the
AWS cluster creation. However at the very end does a dict update with any
custom parameters the user adds. Since this update is not recursive, if
the user sets a nested option that mrjob specifies, then user will override
everything in that nested option. Therefore we must perform a recursive
update.